Make backbone of execution pipeline
Major Change: moved Group and GroupReturn into one file
