+++ src/lodepy/groups/__init__.py
@@ -1,3 +1,3 @@
 'group module imports'
 from lodepy.groups.group import Group
-from lodepy.groups.group_return import GroupReturn
+from lodepy.groups.group import GroupReturn
