+++ src/lodepy/tasks/task.py
@@ -1,13 +1,17 @@
 from abc import abstractmethod
+from typing import TypeVar
 
-from lodepy.nodes.node import Node
+from lodepy.nodes.node_executor import NodeExecutor
+from lodepy.nodes.node_return import NodeReturn
 
-class Task:
+K = TypeVar('K')
+
+class Task: # Note, a task should be able to execute a task using node_executor and then use the NodeReturn for comparisons.
     '''
     An abstract class for Tasks that can be run on Nodes
     '''
     @abstractmethod
-    def execute(self, node: Node) -> None:
+    def execute(self, node_executor: NodeExecutor) -> 'NodeReturn[K]':
         '''
         The method to run when the task gets executed
         '''
